attempt to handle x86_64, where int64_t is typedefed to long at userlevel, but not in linuxmodule.  Also some bug fixes reported by Qinghua Ye

+++ bsdmodule/module.cc
@@ -54,8 +54,8 @@ read_cycles(Element *, void *)
 static String
 read_meminfo(Element *, void *)
 {
-    extern int click_new_count; /* glue.cc */
-    extern int click_outstanding_news; /* glue.cc */
+    extern size_t click_new_count; /* glue.cc */
+    extern size_t click_outstanding_news; /* glue.cc */
     StringAccum sa;
     sa << "outstanding news " << click_outstanding_news << "\n";
     sa << "news " << click_new_count << "\n";
